chemprop2/requirements.txt
flask>=1.1.2
gunicorn>=20.0.4
hyperopt>=0.2.3
matplotlib>=3.1.3
numpy>=1.18.1
pandas>=1.0.3
pandas-flavor>=0.2.0
scikit-learn>=0.22.2.post1
scipy>=1.4.1
tensorboardX>=2.0
tensorflow==1.13.1
torchvision>=0.5.0
tqdm>=4.45.0
typed-argument-parser>=1.4.3
git+https://github.com/bp-kelley/descriptastorus
